Move value-deletion methods to CustomFieldService
delete_custom_field_values no longer only concerns user fields since
field deletion started clearing input answers, so UserCustomFieldService
had become a misnomer. Both its methods are definition-deletion cleanup;
they now live in the general CustomFieldService and the class is gone.
