+++ back/app/services/side_fx_custom_field_service.rb
@@ -14,7 +14,7 @@ class SideFxCustomFieldService
   end
 
   def before_destroy(custom_field, _current_user)
-    UserCustomFieldService.new.delete_custom_field_values(custom_field)
+    CustomFieldService.new.delete_custom_field_values(custom_field)
   end
 
   def after_destroy(frozen_custom_field, current_user)
